pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Nikita Nakleskin <nakleskin.nv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "2.23.1"
python-dotenv = "^1.0.0"
firebase-admin = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
